refactor(spa-editor): split integration-registry entries to fix prettier debt
INTEGRATION_REGISTRY's one-liner entries (0a3021e6) violate Prettier's
print width; running `prettier --write` in place explodes each entry to
a multi-line object, pushing the file over the per-file line cap. Move
the raw entry data to a new integration-registry-entries.ts companion
(same pattern as dexie-schemas-late.ts) and re-export the type + build
INTEGRATION_REGISTRY from it. No behavior change, no public API change
(same exports, same import path for every consumer).
