+++ packages/workout-spa-editor/src/integrations/integration-registry.ts
@@ -20,45 +20,12 @@
 import type { ManagedDataType } from "@kaiord/core";
 import { MANAGED_DATA_REGISTRY } from "@kaiord/core";
 
-import type { ConnectionMechanism } from "../types/connection";
 import type { IntegrationPolicyDirection } from "../types/integration-policy";
+import { INTEGRATION_REGISTRY_ENTRIES } from "./integration-registry-entries";
 
-export type IntegrationRegistryEntry = {
-  id: string;
-  name: string;
-  mark: string;
-  mechanism: ConnectionMechanism;
-  bridgeId: string | null;
-};
+export type { IntegrationRegistryEntry } from "./integration-registry-entries";
 
-export const INTEGRATION_REGISTRY: readonly IntegrationRegistryEntry[] = [
-  { id: "garmin", name: "Garmin", mark: "G", mechanism: "bridge", bridgeId: "garmin-bridge" },
-  { id: "whoop", name: "WHOOP", mark: "Wh", mechanism: "bridge", bridgeId: "whoop-bridge" },
-  {
-    id: "train2go",
-    name: "Train2Go",
-    mark: "T2",
-    mechanism: "bridge",
-    bridgeId: "train2go-bridge",
-  },
-  { id: "manual", name: "Manual Entry", mark: "M", mechanism: "manual", bridgeId: null },
-  {
-    id: "intervals",
-    name: "intervals.icu",
-    mark: "i",
-    mechanism: "api-key",
-    bridgeId: null,
-  },
-  { id: "strava", name: "Strava", mark: "S", mechanism: "not-supported", bridgeId: null },
-  { id: "wahoo", name: "Wahoo", mark: "W", mechanism: "not-supported", bridgeId: null },
-  {
-    id: "trainingpeaks",
-    name: "TrainingPeaks",
-    mark: "TP",
-    mechanism: "not-supported",
-    bridgeId: null,
-  },
-];
+export const INTEGRATION_REGISTRY = INTEGRATION_REGISTRY_ENTRIES;
 
 /** Bridge ids for every "bridge"-mechanism integration. Replaces the two
     independent KNOWN_BRIDGE_IDS lists that used to drift out of sync. */
